src/video_api/requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.78.0
ffmpeg==1.4
ffprobe-python==1.0.3
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
pydantic==1.9.1
redvid==1.2.0
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
